SSL Missing in README.md
As i have pointed out on twitter. This STILL IS a horrible thing to do.
You have NO control on the external domain and there is a possibility that someone could hijack this bit.ly link.\

ATLEAST create an internal scaleway domain for this, so you have a little more control.